Fixed Encoder on Matrix (again)
I got it messed up, A and B should have separate power sources (columns) and common should go to the row. If I had vertical orientation I would have common be on the column and A and B on separate rows. Previous caused a short between columns that could cause ghosting.